chore: move tests folder to internal subfolder to unexport tests
