+++ internal/tests/dummy/dummy.proto
