+++ internal/tests/metadata/metadata.proto
