+++ mgr_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/relab/gorums"
-	"github.com/relab/gorums/tests/dummy"
+	"github.com/relab/gorums/internal/tests/dummy"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 )
